Return array slices.
Return array slices instead of an iterator. Chunking through an iterator
can't be faster than an array slice if the array slice is not long. We
can return 32, 64, 512 at a time. Allows for large pages, but fast
iteration. Eliminates any concern about concurrent writes to the page
because we descend the tree for each slice.

Closes #103.