Correctly implement reverse less than.
When the key is missing we want the first value less than the key but we
were getting the first value greater than the key. This is how a forward
search works. We do get an indication the value was not found with the
`found` boolean and we use that to decrement the index if our
`inclusive` flag is true on reverse. It will only be true on the first
iteration of the query.

Closes #248.