Update for partials in Strata, Reciprocate.
Reciprocate now exports an object with a single `Trampoline` property.

Strata now requires arrays and can perform a search on a partial key so
that if we specify a maximum number of stored key values to compare
against a given key that is longer than that maximum number we're going
to land right after the last value that matched the partial key, or
where that last value would be if it existed.